Moved some more of the methods from the abstract class back to the n2 gui test file because of the use of the global test counter
